Updateded flag behaviour
All flags now repair 1 hp when you stand on one at the end of a phase, even tho it is not the correct flag in order. Repairs is the only thing that changes the backup of a robot.
That means backup will be relocated every time the robot is healed, no exceptions.